acp: extract ACPRuntimeSupervisor base; migrate Hermes to subclass it
The generic ACP turn lifecycle (start/wait/stream/interrupt, event buffering,
session create/resume/list/fork, and the approval flow) was duplicated inside
HermesSupervisor. Promote it to a shared agents/acp/runtime_supervisor.py base
(ACPRuntimeSupervisor) parameterized by runtime_id, runtime_label, error_cls,
env, and an overridable approval-policy hook. ACPSubprocessSupervisor now also
threads the G1 missing_session_matcher through to ACPClient.

HermesSupervisor becomes a thin subclass supplying only Hermes specifics
(profile-PATH env, runtime label/error type, binary resolution + preflight).
HermesSupervisorError now subclasses ACPRuntimeSupervisorError and keeps the
exact turn-error messages; HermesSessionHandle aliases the shared ACPSessionHandle.
registry.py migrated to the generic ApprovalHandler.

Verified: 250 acp+hermes tests pass; mypy strict clean (997 files).
